project/requirements/common.txt
celery[redis]==4.3.0
kombu==4.6.3
amqp==2.5.0
billiard==3.6.0.0
Django==2.2.4
django-constance[database]==2.4.0
django-cors-headers==3.0.2
django-environ==0.4.5
django-fieldsignals==0.4.0
django-filter==2.2.0
django-guardian==2.0.0
django-phonenumber-field==3.0.1
django-redis==4.10.0
django-watchman==0.17.0
drf-yasg==1.16.1
djangorestframework==3.10.2
djangorestframework-guardian==0.3.0
djangorestframework-recaptcha==0.2.0
djangorestframework-jwt==1.11.0
django-fsm==2.6.1
gunicorn==19.9.0
phonenumbers==8.10.16
psycopg2-binary==2.8.3
pytz==2019.2
readline==6.2.4.1
redis==3.3.5
sentry-sdk==0.10.2
vine==1.3.0
whitenoise==4.1.3
